requirements/develop.txt
# Everything the developer needs in addition to the production requirements
-r prod.txt

# Testing
pytest==3.3.1
WebTest==2.0.29
factory-boy==2.9.2

# Lint and code style
flake8==3.5.0
flake8-blind-except==0.1.1
flake8-debugger==3.0.0
flake8-docstrings==1.1.0
flake8-isort==2.3
flake8-quotes==0.13.0
isort==4.2.15
pep8-naming==0.4.1

flask-shell-ipython==0.3.0
flask-sqlacodegen==1.1.6.1
ipython==5.5.0
